fix(DTO): Fix DTO to separate old version with new version
